(NPUP-39) Fix type hashing.
The `hash_value` implementation for several types were accidentally hashing
pointer values instead of hashing the types they reference.  This commit fixes
that such that the pointers are properly dereferenced.

Iterable and Iterator did not correctly implement operator== such that it
always returned true when the types were specialized.

Callable did not correctly implement operator== such that comparing two
Callables that did not have block types resulted in a segfault (dereference of
null block type).  This commit fixes that logic.

Also included is a test that hashes every type and ensures the hash can be
accessed again with the same type.