(NPUP-38) Implement 'environment' scope.
This commit implements the 'environment' scope.

Note: 'environment' is no longer valid as as an environment name and will error
if used.

Functions and types can now be loaded with a namespace of 'environment'. This
will load the files from the current environment.

This commit also adds test coverage for auto loading manifests, types, and
functions from an environment and modules.