(maint) Fix Struct type output.
The `Struct` type was not outputting the contents of the schema as a hash,
resulting in the `Struct` appearing to take multiple arguments when it only
accepts one argument (the schema hash).  This commit fixes this by outputting
the matching `{` and `}` around the argument.

Additionally, string keys would always show up as `Enum[xyz]`.  This commit
fixes that to output enumerations as just the string value.