(NPUP-41) Implement the `type` function.
This commit implements the `type` function and modifies the parser to allow
the type keyword as a function name.

Also implemented is the "detailed" and "generalized" forms of type inference.

For the "generalized" form, each type now implements a `generalized` member
function that returns the same type, but with default size constraints.

For the "detailed" form, the `infer_type` member function was updated to handle
arrays and hashes so that a `Tuple`, `Struct`, or `Hash` type is returned that
fully describes the contents of the array or hash.