(NPUP-28) Integrate Onigmo regular expression library.
This commit integrates the Onigmo regular expression library.  The library is
the same one used by Ruby for its regular expression support.  We're using it
to provide maximum compatibility with regular expression syntax and options.

This commit also contains tests for the `split` and `versioncmp` functions
because they use regular expression iterators for their implementation.