(NPUP-42) Implement new expression.
This commit implements support in the parser and evaluator for a 'new
expression', which allows calls on types to evaluate to a "new" instance of
that type.

The evaluator translates the 'new expression' into a call to a `new` function, which
is not yet implemented.

The implementation of the `new` function and corresponding tests will be
included in a forthcoming commit.