(maint) Move parser tests out of 'good' directory.
Originally the parser tests were implemented such that we needed to separate
out 'good' baselines from 'bad' (those with inputs that cause exceptions)
baselines.  When the evaluation tests were implemented, they were implemented
such that it was possible for good and bad to live side-by-side, with the
baseline themselves containing any compiler output.

This commit fixes the parser tests so that they are implemented in the same way
as the evaluation tests.

The input files and their baselines have moved up one directory as a result;
additionally, the baselines now contain a 'normalized' path in the AST output.