(NPUP-43) Fix resource body rule to expect ':' following title expression.
The resource body rule was not expecting `:` following the title expression,
leading to a confusing error message when given `file { foo }` as input, where
the colon was missing.

This fixes the rule such that it will expect `:` following the title, provided
the token isn't `=>`, which is needed to match other rules.