(maint) Fix binary operator evaluation AST context.
Previously, the AST context of an evaluated subexpression was not being
adjusted so error messages might highlight only part of the subexpression (the
left-hand side).

For example, evaluating `1 + 2 + 3 in foo` results in an error because the `in`
operator has precedence, resulting in the addition of an integer and a boolean.
Previously, the error message would only point at the `3` as the problem.

With this commit, the AST context is updated to include the RHS.  Thus, with
the above example, the error message now points at the entire `3 in foo`
subexpression.