(NPUP-44) Fix parser rules to assist AST validation.
The parser rules were previously defined to be too flexible, causing a number
of invalid expressions to parse which would later have to be caught during a
validation pass.

Instead of implementing those validations, this commit refactors the parser rules
and AST so that the parser itself will catch invalid syntax.  There are now
formal statements in the grammar:

* Definition statements (class, defined type, node, function, produces,
  consumes, application, site, and type aliases).
* Function call statement (similar to a function call expression, except has no
  parenthesis, only works for a defined set of functions, and cannot appear as
  part of an expression).
* Relationship statements.

Whereby a relationship statement is defined to be an expression followed by
zero or more relationship operations.  A relationship statement with no
operations is simply a "general" expression; it was done this way to prevent
having the parser backtrack after not finding a relationship operator following
an expression.

These new rules eliminate the need for "rvalue" validation which is done in the
Ruby parser because operands to the binary operators are always
value-producing.  Subsequently, collectors can only be used in relationship
statements and in resource override expressions based on the updated rules.

More AST validations will be implemented in upcoming commits.