(NPUP-48) Implement finding files and templates in environments/modules.
This commit implements the resolving of files and templates by path within
environments and modules.

This functionality is needed to implement the upcoming `file` and `epp`
functions.

Also included is a change to the evaluation tests to run them in an
'evaluation' environment that is local to the test fixtures directory.  This
will be used to test the `file` and `epp` functions.