Implement a C API for simple Puppet file evaluation.
This commit implements a C API for evaluating a Puppet source file and
returning a C representation of a Puppet runtime value.

Consumers of the API can hook into logged messages and evaluate individual
source files, getting the resulting value back.

An API for creating Puppet values is also implemented for the future feature of
using this API for defining Puppet functions.