Fix evaluation exception to use the nearest source context.
When constructing an evaluation exception from a backtrace without source
context, no source context was used for displaying error messages.

This commit changes this behavior such that the top frame with context is used
as the source context.  Thus when external (i.e. native) frames are at the top
of the backtrace, the nearest Puppet frame is used for source context.