(NPUP-49) Implement break statement in the Puppet language.
This commit implements the `break` statement in the Puppet language.

Unlike in the Ruby implementation, `break` is not implemented as a function
that uses exception handling to implement non-local control transfer.

In the native compiler's implementation, `break` is implemented as a language
statement and static analysis is performed to prevent illegal uses.  Instead of
using exception handling, a special value is evaluated for the `break`
statement which must be respected by iterating functions.  Failure to respect
the statement will result in an evaluation exception.

Also included were a few bug fixes the new tests caught.