(NPUP-50) Implement the next statement.
This commit implements the next statement in the Puppet language.

Next statements are only allowed inside of lambdas and function like a "yield
return", that is to say a return from the current block instead of the "return"
statement that returns from the current function.

Like the break statement, this is implemented without exception handling for
performance reasons.